+++ packages/devtools_app/lib/src/shared/banner_messages.dart
@@ -3,10 +3,8 @@
 // found in the LICENSE file.
 
 import 'package:collection/collection.dart' show IterableExtension;
-import 'package:devtools_app_shared/service.dart';
 import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_app_shared/utils.dart';
-import 'package:devtools_shared/devtools_shared.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
@@ -481,61 +479,6 @@ For the most accurate absolute memory stats, relaunch your application in ''',
         ),
       ],
       screenId: screenId,
-    );
-  }
-}
-
-class UnsupportedFlutterVersionWarning {
-  const UnsupportedFlutterVersionWarning({
-    required this.screenId,
-    required this.currentFlutterVersion,
-    required this.supportedFlutterVersion,
-  });
-
-  final String screenId;
-
-  final FlutterVersion currentFlutterVersion;
-
-  final SemanticVersion supportedFlutterVersion;
-
-  BannerMessage build() {
-    return BannerWarning(
-      key: Key('UnsupportedFlutterVersionWarning - $screenId'),
-      textSpans: [
-        TextSpan(
-          text: 'This version of DevTools expects the connected app to be run'
-              ' on Flutter >= $supportedFlutterVersion, but the connected app'
-              ' is running on Flutter $currentFlutterVersion. Some'
-              ' functionality may not work. If this causes issues, try'
-              ' upgrading your Flutter version.',
-          style: TextStyle(fontSize: defaultFontSize),
-        ),
-      ],
-      screenId: screenId,
-    );
-  }
-}
-
-void maybePushUnsupportedFlutterVersionWarning(
-  String screenId, {
-  required SemanticVersion supportedFlutterVersion,
-}) {
-  final isFlutterApp =
-      serviceConnection.serviceManager.connectedApp?.isFlutterAppNow;
-  if (offlineController.offlineMode.value ||
-      isFlutterApp == null ||
-      !isFlutterApp) {
-    return;
-  }
-  final currentVersion =
-      serviceConnection.serviceManager.connectedApp!.flutterVersionNow!;
-  if (currentVersion < supportedFlutterVersion) {
-    bannerMessages.addMessage(
-      UnsupportedFlutterVersionWarning(
-        screenId: screenId,
-        currentFlutterVersion: currentVersion,
-        supportedFlutterVersion: supportedFlutterVersion,
-      ).build(),
     );
   }
 }