+++ packages/devtools_shared/CHANGELOG.md
@@ -1,5 +1,4 @@
 # 8.0.0
-* Refactor yaml extension methods.
 * **Breaking change:** rename `ServerApi.getCompleted` to `ServerApi.success` and make the
 `value` parameter optional.
 * **Breaking change:** remove the `String? dtdUri` parameter from `ServerApi.handle` and replace
@@ -7,6 +6,8 @@ it with a parameter `DTDConnectionInfo? dtd`.
 * Introduce a new typedef `DTDConnectionInfo`.
 * Add a new API `apiNotifyForVmServiceConnection` that DevTools will call when a
 VM service connection is connected or disconnected from the client.
+* Refactor yaml extension methods.
+* Add a helper method `packageRootFromFileUriString`.
 
 # 7.0.0
 * **Breaking change:** remove the `ServerApi.setCompleted` method that was a